WIP move types for Octokit class into @octokit/types
