+++ packages/types-rest-api-ghes-3.0/index.d.ts
@@ -1,6 +1,6 @@
 import { Octokit } from "@octokit-next/types";
 
-import "@octokit-next/types-rest-api-3.1";
+import "@octokit-next/types-rest-api-ghes-3.1";
 
 type GHES30ResponseHeaders = Octokit.ApiVersions["ghes-3.1"]["ResponseHeaders"];
