+++ tests/js/ghes-3.0/index.js
