+++ tests/js/plugins-and-defaults/index.d.ts
