+++ tests/js/plugins-and-defaults/index.js
