+++ tests/js/plugins-and-defaults/index.test-d.ts
