+++ tests/js/plugins-and-defaults/plugins/bar/index.d.ts
@@ -1,4 +1,4 @@
-import { Octokit } from "../../../../index.js";
+import { Octokit } from "@octokit-next/types";
 
 export function barPlugin(base: Octokit): {
   bar: string;
