+++ tests/js/plugins-and-defaults/plugins/bar/index.js
