+++ tests/js/plugins-and-defaults/plugins/foo/index.d.ts
@@ -1,4 +1,4 @@
-import { Octokit } from "../../../../index.js";
+import { Octokit } from "@octokit-next/types";
 
 export function fooPlugin(
   base: Octokit,
