+++ tests/js/plugins-and-defaults/plugins/foo/index.js
