+++ tests/js/plugins-and-defaults/plugins/void/index.d.ts
@@ -1,3 +1,3 @@
-import { Octokit } from "../../../../index.js";
+import { Octokit } from "@octokit-next/types";
 
 export function voidPlugin(base: Octokit, options: Octokit.Options): void;
