+++ tests/js/plugins-and-defaults/plugins/void/index.js
